% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeMissValues.R
\name{imputeMissValues}
\alias{imputeMissValues}
\title{Impute missing covariate values}
\usage{
imputeMissValues(data = NULL, covariates = NULL,
  imputation_metrics = NULL, max_miss = 3, response = NA)
}
\arguments{
\item{data}{data.frame containing the response metric, the covariates and any other imputation metrics}

\item{covariates}{character vector of names of covariates to be used in QRF model}

\item{imputation_metrics}{character vector of names of additional covariates to be used in imputing missing values in the \code{covariates}.}

\item{max_miss}{maximum number of missing covariate values for any given data point. Data points with more than this number of missing values will be deleted.}

\item{response}{column name of response value}
}
\value{
data.frame with response vector in first column and covariates in subsequent columns.
}
\description{
filter rows with too many missing values, and imputes the remaining missing values.
}
\examples{
imputeMissValues(data = mod_data, covariates = hab_mets, imputation_metrics = impute_mets, response = 'chnk_per_m')
}
\author{
Kevin See
}

