% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relImportance.R
\name{relImportance}
\alias{relImportance}
\title{Calculate relative importance of covariates}
\usage{
relImportance(qrf_mod)
}
\arguments{
\item{qrf_mod}{object of class \code{quantregForest} returned by the function \code{quantregForest}}
}
\description{
Calculate node purity and relative importance
}
\examples{
relImportance(qrf_mod)
}
\author{
Kevin See
}

