% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMine.R
\name{runMINE}
\alias{runMINE}
\title{QRF Variable Selection}
\usage{
runMINE(data = NULL, covariates = NULL, response = NULL,
  covar_dict = NULL, save_csv_file = F)
}
\arguments{
\item{data}{\code{data.frame} that was used to fit the \code{mod}}

\item{covariates}{character vector of names of possible covariates}

\item{response}{column name of response value}

\item{covar_dict}{\code{data.frame} containing columns called ShortName, Name, and MetricCategory. ShortName corresponds to the covariates that went into the \code{mod}, while Name is a longer version for plotting purposes. MetricCategory is a grouping factor}

\item{save_csv_file}{should the results be written to a .csv file?}
}
\description{
Calculates the maximal information criteria for all pairwise groups of the response metric and covariates.
}
\author{
Kevin See
}

