% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relImpPlot.R
\name{relImpPlot}
\alias{relImpPlot}
\title{Plot relative importance of covariates}
\usage{
relImpPlot(mod, covar_dict, fill = "gray40", title = "Relative Importance",
  relImp_var = c("relImp", "IncNodePurity"))
}
\arguments{
\item{mod}{object of class \code{quantregForest} returned by the function \code{quantregForest}}

\item{covar_dict}{\code{data.frame} containing columns called ShortName and Name. ShortName corresponds to the covariates that went into the \code{mod}, while Name is a longer version for plotting purposes.}
}
\description{
plot the relative importance of each covariate in a quantile regression forest model
}
\examples{
relImpPlot(qrf_mod)
}
\author{
Kevin See
}

