% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMINE.R
\name{runMINE}
\alias{runMINE}
\title{QRF Variable Selection - MINE}
\usage{
runMINE(data = NULL, covariates = NULL, response = NULL,
  covar_dict = NULL, save_csv_file = F)
}
\arguments{
\item{data}{\code{data.frame} that was used to fit the \code{mod}}

\item{covariates}{character vector of names of possible covariates}

\item{response}{column name of response value}

\item{covar_dict}{\code{data.frame} containing columns called ShortName, Name, and MetricCategory. ShortName corresponds to the covariates that went into the \code{mod}, while Name is a longer version for plotting purposes. MetricCategory is a grouping factor}

\item{save_csv_file}{should the results be written to a .csv file?}
}
\description{
Calculates the maximal information criteria for all pairwise groups of the response metric and covariates.
}
\examples{
varSelect_df = all_data_clean \%>\%
filter(chnk_per_m > 0) \%>\%
  select(chnk_per_m,
         one_of(as.character(hab_dict$ShortName[hab_dict$ShortName \%in\% names(all_data_clean)]))) \%>\%
  select(-one_of(as.character(hab_dict$ShortName[hab_dict$MetricCategory == 'Categorical'])))
runMINE(varSelect_df, names(varSelect_df)[-1], names(varSelect_df)[1], hab_dict)
}
\author{
Kevin See
}
\references{
Reshef, D. N., Y. A. Reshef, H. K. Finucane, S. R. Grossman, G. McVean, P. J. Turnbaugh, E. S. Lander, M. Mitzenmacher, and P. C. Sabeti (2011). Detecting novel associations in large data sets. Science, 334(6062):1518–1524.
}

